requirements.txt
datasets
openai
python-dotenv